DOMA-3400 moved getTicketAttributesFilter tests to ticket domain
