+++ apps/condo/domains/common/utils/tables.utils.spec.ts
@@ -13,7 +13,7 @@ import {
     getFiltersFromQuery,
     getPageIndexFromOffset,
     getSorterMap,
-    convertColumns, getTicketAttributesFilter,
+    convertColumns,
 } from './tables.utils'
 import dayjs from 'dayjs'
 
@@ -278,39 +278,6 @@ describe('Table utils', () => {
         })
         it('should accept only single argument', () => {
             const result = filter(['true', 'false'])
-            expect(result).toBeUndefined()
-        })
-    })
-    describe('getTicketAttributesFilter', () => {
-        const isEmergencyAttribute = 'isEmergency'
-        const isPaidAttribute = 'isPaid'
-        const isWarrantyAttribute = 'isWarranty'
-        const filterAttribute = getTicketAttributesFilter([isEmergencyAttribute, isPaidAttribute, isWarrantyAttribute])
-
-        it('should return filter query with one search argument', () => {
-            const result = filterAttribute(isEmergencyAttribute)
-
-            expect(result).toBeDefined()
-            expect(result).toStrictEqual({ OR: [ { [isEmergencyAttribute]: true } ] })
-        })
-
-        it('should return filter query with two search argument', () => {
-            const result = filterAttribute([isEmergencyAttribute, isPaidAttribute])
-
-            expect(result).toBeDefined()
-            expect(result).toStrictEqual({ OR: [ { [isEmergencyAttribute]: true }, { [isPaidAttribute]: true } ] })
-        })
-
-        it('should return filter query with three search argument', () => {
-            const result = filterAttribute([isEmergencyAttribute, isPaidAttribute, isWarrantyAttribute])
-
-            expect(result).toBeDefined()
-            expect(result).toStrictEqual({ OR: [ { [isEmergencyAttribute]: true }, { [isPaidAttribute]: true }, { [isWarrantyAttribute]: true } ] })
-        })
-
-        it('should return undefined with wrong search argument', () => {
-            const result = filterAttribute('')
-
             expect(result).toBeUndefined()
         })
     })
